chore: create select helpers
